+++ src/librustc/middle/trans/adt.rs
@@ -282,13 +282,13 @@ fn load_discr(bcx: block, scrutinee: ValueRef, min: Disr, max: Disr)
 pub fn trans_case(bcx: block, r: &Repr, discr: Disr) -> _match::opt_result {
     match *r {
         CEnum(*) => {
-            _match::single_result(rslt(bcx, C_int(bcx.ccx(), discr /*bad*/as int)))
+            _match::single_result(rslt(bcx, C_int(bcx.ccx(), discr)))
         }
         Univariant(*)=> {
             bcx.ccx().sess.bug(~"no cases for univariants or structs")
         }
         General(*) => {
-            _match::single_result(rslt(bcx, C_int(bcx.ccx(), discr /*bad*/as int)))
+            _match::single_result(rslt(bcx, C_int(bcx.ccx(), discr)))
         }
     }
 }
@@ -302,7 +302,7 @@ pub fn trans_start_init(bcx: block, r: &Repr, val: ValueRef, discr: Disr) {
     match *r {
         CEnum(min, max) => {
             fail_unless!(min <= discr && discr <= max);
-            Store(bcx, C_int(bcx.ccx(), discr /*bad*/as int), GEPi(bcx, val, [0, 0]))
+            Store(bcx, C_int(bcx.ccx(), discr), GEPi(bcx, val, [0, 0]))
         }
         Univariant(ref st, true) => {
             fail_unless!(discr == 0);
@@ -313,7 +313,7 @@ pub fn trans_start_init(bcx: block, r: &Repr, val: ValueRef, discr: Disr) {
             fail_unless!(discr == 0);
         }
         General(*) => {
-            Store(bcx, C_int(bcx.ccx(), discr /*bad*/as int), GEPi(bcx, val, [0, 0]))
+            Store(bcx, C_int(bcx.ccx(), discr), GEPi(bcx, val, [0, 0]))
         }
     }
 }
@@ -404,7 +404,7 @@ pub fn trans_const(ccx: @CrateContext, r: &Repr, discr: Disr,
         CEnum(min, max) => {
             fail_unless!(vals.len() == 0);
             fail_unless!(min <= discr && discr <= max);
-            C_int(ccx, discr /*bad*/as int)
+            C_int(ccx, discr)
         }
         Univariant(ref st, _dro) => {
             fail_unless!(discr == 0);
@@ -414,7 +414,7 @@ pub fn trans_const(ccx: @CrateContext, r: &Repr, discr: Disr,
             let case = &cases[discr as uint];
             let max_sz = cases.map(|s| s.size).max();
             let contents = build_const_struct(ccx, case,
-                                              ~[C_int(ccx, discr/*bad*/ as int)] + vals);
+                                              ~[C_int(ccx, discr)] + vals);
             C_struct(contents + [padding(max_sz - case.size)])
         }
     }
